+++ NEWS.md
@@ -1,9 +1,18 @@
+## R package processR version 0.1.8
+===================================
+(2019-Apr-9)
+
+- New functions modelsSummary2, print.modelSummary2, modelsSummary2Table added
+
+- New functions compareMC, compareMCTable, fit2vif, compareVIF, print.compareVIF and compareVIFTable added
+
 ## R package processR version 0.1.7
 ===================================
 (2019-Apr-8)
 
 - Bug- fixed
 
+- New functions standardize, standardizeDf added
 
 ## R package processR version 0.1.6
 ===================================
@@ -17,20 +26,20 @@
 ===================================
 (2019-Mar-29)
 
-- New function mediationBK, print.mediationBK, summary.mediationBK and plot.mediationBK added
+- New functions mediationBK, print.mediationBK, summary.mediationBK and plot.mediationBK added
 
 
 ## R package processR version 0.1.4
 ===================================
 (2019-Mar-26)
 
-- New function medSummary and medSummaryTable added
+- New functions medSummary and medSummaryTable added
 
 ## R package processR version 0.1.3
 ===================================
 (2019-Mar-22)
 
-- New function extractRange and extractnumber added
+- New functions extractRange and extractnumber added
 
 
 ## R package processR version 0.1.2